MatrixPilot_AUAV3 branch: moving to generic microchip USB/MDD library structure, work in progress...
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_AUAV3@2460 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ libFlashFS/minIni.c
@@ -827,7 +827,7 @@ int ini_putl(const TCHAR *Section, const TCHAR *Key, long Value, const TCHAR *Fi
 int ini_putf(const TCHAR *Section, const TCHAR *Key, INI_REAL Value, const TCHAR *Filename)
 {
   TCHAR LocalBuffer[64];
-  ini_ftoa(LocalBuffer, Value);
+  ini_ftoa(LocalBuffer, (double)Value);
   return ini_puts(Section, Key, LocalBuffer, Filename);
 }
 #endif /* INI_REAL */
